requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
llama-cpp-python>=0.2.0
huggingface-hub>=0.19.0

